fix: address code review findings in mission stats feature
Fix concurrency bugs: add ConcurrentMessageLimit(1) on stats consumer,
replace read-modify-replace with atomic MongoDB $inc operations, combine
session updates into single atomic update. Fix design issues: change
MissionSessionsContext from CachedMongoContext to MongoContext, add
MongoDB indexes for all stats collections, use AsInt() extension for
variable access with debug logging on failure, change TotalDistance from
int to double. Optimize: parallelize per-player DB writes with
Task.WhenAll, pre-allocate event list capacity. Add tests for
server_status/performance event handlers, empty events, mock processors
in consumer tests.

+++ UKSF.Api.ArmaServer/Models/PlayerMissionStats.cs
@@ -17,5 +17,5 @@ public class PlayerMissionStats : MongoObject
     public int TotalHits { get; set; }
     public Dictionary<string, WeaponStats> WeaponBreakdown { get; set; } = new();
     public Dictionary<string, int> BodyPartHits { get; set; } = new();
-    public int TotalDistance { get; set; }
+    public double TotalDistance { get; set; }
 }
